Add more types and fmt::print to testfmt
This commit adds in tests for several different data types using
fmt::format_to and sprintf. Each of these tests is timed using chrono's
steady_clock, a monotonic clock with nanosecond precision. With this, we
are able to determine that when compiled in debug mode (meaning that
fmt is added as a shared library), the runtime for the formatting is
about 1000 nanoseconds with fmt::format_towhile sprintf is about 200
nanoseconds. In release mode CMake's standard release modeflags (I
believe this includes -O3), both run in about 180 nanoseconds. These
averaages are both calculated on an Intel Core i7-8750H CPU (12 threads
@ 2.20GHz base, 4.10 GHz turbo) after one million formatting iterations.